This is a Devolping AI CHATBOT
This can answer any questions because this have access to google through a GOOGLE CUSTOM SEARCH ENGINE & API. So this can answer any question. First you have to sign up with a given form in the file and you can chat with it anything. It can even understand foul words because this can aviod the inappropriate conversation between user and the chatbot